+++ grails-app/controllers/newstore/UrlMappings.groovy
@@ -8,6 +8,8 @@ class UrlMappings {
                 // apply constraints here
             }
         }
+        "/search"(controller: 'search', action: 'search')
+
         "/"(controller: 'meal', action: 'index')
         "500"(view:'/error')
         "404"(view:'/notFound')
