+++ src/test/groovy/newstore/SearchControllerSpec.groovy
@@ -1,0 +1,19 @@
+package newstore
+
+import MealStore.SearchController
+import grails.testing.web.controllers.ControllerUnitTest
+import spock.lang.Specification
+
+class SearchControllerSpec extends Specification implements ControllerUnitTest<SearchController> {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
